# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Koishi API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://koishi-api.nechapu.to",
      "description": "The official Koishi API server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Koishi API running on localhost"
    }
  ],
  "components": {
    "schemas": {
      "ObjectId": {
        "type": "string",
        "example": "625401f6ffd69c32c336e0ee"
      },
      "x-auth-token": {
        "type": "string",
        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MjBmOWNjYzc4NWFkMjNhNTkwNWJkOTgiLCJuYW1lIjoiSG9uemEiLCJlbWFpbCI6ImhvbnphQG5lY2hhcHUudG8iLCJzdGF0dXMiOiJ2ZXJpZmllZCIsImlzQWRtaW4iOnRydWUsImlhdCI6MTY0OTc1MzU0OH0.loyrUX_qXMLG5hyi9_Vcudst36KsEIqMA8gZzIRAEwA"
      },
...
      "User": {
        "type": "array",
        "description": "User data the way it is stored in the DB",
        "items": {
          "type": "object",
          "properties": {
            "_id": {
              "$ref": "#/components/schemas/ObjectId"
            },
            "name": {
              "type": "string",
              "example": "Hayley"
            },
            "email": {
              "type": "string",
              "example": "hayley@gmail.com"
            },
            "status": {
              "type": "string",
              "enum": ["pending", "verified"]
            },
            "isAdmin": {
              "type": "boolean"
            },
            "verificationCode": {
              "type": "string",
              "example": "c2775912ebc49c4211365d25b50f90f2"
            },
            "collections": {
              "type": "array",
              "items": {
                "description": "ObjectIDs referencing all the user's collections, including those that the user accepted share invite for.",
                "$ref": "#/components/schemas/ObjectId"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "authToken": {
        "type": "apiKey",
        "name": "x-auth-token",
        "in": "header"
      }
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "description": "Get all registered users.",
        "security": [
          {
            "authToken": []
          }
        ],
        
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized. - meaning your token does not have the admin access level."
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "description": "Register a new user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Pam"
                  },
                  "email": {
                    "type": "string",
                    "example": "pamela@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "NdawAS9sA62"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully registered, returns a x-auth-token",
            "content": {
              "plain/text": {
                "schema": {
                  "$ref": "#/components/schemas/x-auth-token"
                }
              }
            }
          },
          "400": {
            "description": "This email is already registered"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["Users"],
        "description": "Get information about the user that is encoded in the provided x-auth-token",
        "security": [
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/verify": {
      "post": {
        "tags": ["Users"],
        "description": "If the email verification is enabled, this endpoint is used to verify a newly registered user.",
        "parameters": [
          {
            "description": "The verification token (is part of an url provided in auto generated verification emal.",
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "612ace0f91f86cf37a3d94a84af1f543"
            }
          }
        ],
        "security": [
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Verified - returns updated JWT token (the old one has to be replaced, still shows the user status as \"pending\")"
          },
          "400": {
            "description": "Invalid token. / User has already been verified."
          }
        }
      }
    },
...
  }
}